ombudsman_core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ombudsman_core"
version = "0.1.0"
description = "Ombudsman Migration Testing Framework Core"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0.0",
    "pyodbc>=5.0.0",
    "snowflake-connector-python>=3.6.0",
]

[tool.setuptools.packages.find]
where = ["src"]
